Cleanup before writing the tests
