+++ DataReader/ViewController.swift
@@ -20,15 +20,7 @@ class ViewController: UITableViewController {
         super.viewDidLoad()
         tableView.tableFooterView = UIView()
         
-        let notificationCenter = NotificationCenter.default
-        notificationCenter.addObserver(self,
-                                       selector: #selector(didEnterBackground),
-                                       name: NSNotification.Name.UIApplicationWillResignActive,
-                                       object: nil)
-        notificationCenter.addObserver(self,
-                                       selector: #selector(didBecomeActive),
-                                       name: NSNotification.Name.UIApplicationDidBecomeActive,
-                                       object: nil)
+        enableDataChangeFeature()
     }
 
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
@@ -44,36 +36,5 @@ class ViewController: UITableViewController {
         }
 
         return cell
-    }
-    
-    func startTimer() {
-        timer = Timer.scheduledTimer(timeInterval: 1,
-                                     target: self,
-                                     selector: #selector(updateRandomCell),
-                                     userInfo: nil,
-                                     repeats: true)
-    }
-    
-    @objc func didEnterBackground() {
-        timer?.invalidate()
-        timer = nil
-    }
-
-    @objc func didBecomeActive() {
-        startTimer()
-    }
-
-    @objc func updateRandomCell() {
-        let randomRow = Int(arc4random_uniform(UInt32(textObjects.count)))
-        
-        guard let cell = tableView.cellForRow(at: IndexPath(item: randomRow, section: 0)) as? DataTextViewCell else {
-            return
-        }
-        
-        textObjects[randomRow] = String.randomSentence()
-        
-        tableView.beginUpdates()
-        cell.dataTextView.text = textObjects[randomRow]
-        tableView.endUpdates()
     }
 }
