+++ DataReaderTests/ViewControllerTest.swift
@@ -1,0 +1,9 @@
+//
+//  ViewControllerTest.swift
+//  DataReaderTests
+//
+//  Created by Damian Stanczyk on 01.11.2017.
+//  Copyright © 2017 haze. All rights reserved.
+//
+
+import Foundation
